put the filter in single-precision float
An experiment, to be A/B'd against main on bench_max before it is
adopted or dropped.

SVF::apply drops from 34 instructions a sample to 15, with no calls
and no spills: three vfma, one vnmul, one vsub, two conversions and
the loop.  Whether that is actually faster is the open question -
vfma has around three cycles of latency against one of throughput,
and the SVF is a strict serial recurrence, low to high to band to the
next low.  Instruction count is not cycle count here, which is why
this wants measuring rather than arguing about.

The tables stay integer.  Their three coefficients convert once per
buffer, in the preamble, and live in FP registers across the loop -
float tables would save three instructions per buffer and cost twice
the flash.

Numerically it is a clear improvement: against a double precision
reference with the same output clip, SNR goes from 89-92 dB to
114-126 dB, and is bit exact at low levels.  24 bits of mantissa
against the 18 the state occupies, and IEEE rounding makes the
round-to-nearest care that the fixed point version needed automatic.

Two details that are easy to get wrong.  lrintf does not inline - it
has to set errno, so the compiler emits a call, per sample - hence the
vcvtr written out longhand, which also rounds to nearest where a cast
would truncate.  And the FPU is left in flush-to-zero mode on core 1:
with silence at the input the state reaches denormals within 20 ms,
and those are slower than normal operands.  The oscillator never
actually goes silent while a voice exists, so this is insurance.

+++ src/filter.h
@@ -27,7 +27,7 @@ public:
 class SVF : public Filter {
 
 private:
-	int32_t					low, band;
+	float					low, band;
 
 public:
 	virtual void			apply(int16_t* buf, size_t n);
